# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: ""
  title: Visa Account Updater
  version: "1"
servers:
- description: Sandbox server
  url: https://sandbox.api.visa.com
...
paths:
...
  /vau/issuer-api/v1/stopadvices:
    post:
      description: Issuer can place stop advice by providing following requested information
      operationId: stopAdviceUsingPOST
      parameters:
      - explode: false
        in: header
        name: Accept
        required: false
        schema:
          enum:
          - application/json
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/stopAdviceRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopAdviceResponse'
          description: "This response code will be returned in scenarios where the\
            \ stop advice is successful, all possible response codes are explained\
            \ below in the model."
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadFormattedErrorMessage'
          description: This response code will be returned in scenarios where the
            stop advice is badly formatted.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RejectedStopAdviceRequestResponse'
          description: "This response code will be returned in scenarios where the\
            \ stop advice is rejected, all possible reasons for request rejection\
            \ are explained below in the model."
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorMessage'
          description: This response code will be returned in scenarios where the
            stop advice could not be sucessfully processed.
      summary: Issuer Stop Advice API
      tags:
      - Issuer Stop Advice API
      x-operationVersions:
      - label: v1 - Latest
        operationPointer: '#/paths/~1vau~1issuer-api~1v1~1stopadvices/post'
        default: false
      x-hideTryIt: true
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Issuer Stop Advice API
...
  /vau/acquirer-api/v1/inquiry:
    post:
      deprecated: false
      description: Retrieve the new PANs information for requested old PANs.
      operationId: inquiryUsingPOST
      requestBody:
        content:
          application/json:
            examples:
              Default1:
                summary: Default1
                value:
                  inquiries:
                  - cardholderAccountNumber: a4X3933787254355
                    expirationDate: "2609"
                  acquirerSegmentId: 2
                  merchantId: 00000067625P
                  acquirerOrMerchantProprietaryInfo: ""
                  subMerchantName: ""
              Default:
                summary: Default
                value:
                  inquiries:
                  - cardholderAccountNumber: 4389613X6834X49X
                    expirationDate: "1802"
                  - cardholderAccountNumber: 4389A13X6834X49X
                    expirationDate: "1802"
                  acquirerSegmentId: 88
                  merchantId: "000000705858"
                  acquirerOrMerchantProprietaryInfo: ""
                  subMerchantName: SM01
            schema:
              $ref: '#/components/schemas/Request'
        description: inquiryRequest
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                Default1:
                  summary: Default1
                  value:
                    rejects:
                    - cardholderAccountNumber: a4X3933787254355
                      rejectReason: Account Number contains non-numeric characters
                        or is blank
                      rejectReasonCode: "3"
                      expirationDate: "2609"
                    acquirerSegmentId: 2
                    merchantId: 00000067625P
                    responses: []
                    acquirerOrMerchantProprietaryInfo: ""
                    subMerchantName: ""
                Default:
                  summary: Default
                  value:
                    rejects:
                    - cardholderAccountNumber: 4389A13X6834X49X
                      rejectReason: Account Number contains non-numeric characters
                        or is blank
                      rejectReasonCode: "3"
                      expirationDate: "1802"
                    acquirerSegmentId: 88
                    merchantId: "000000705858"
                    responses:
                    - previouslySentFlag: "N"
                      newCardholderAccountNumber: ""
                      oldExpirationDate: "1802"
                      newExpirationDate: ""
                      oldCardholderAccountNumber: 4389613X6834X49X
                      serviceIdentifier: P
                    acquirerOrMerchantProprietaryInfo: ""
                    subMerchantName: SM01
              schema:
                $ref: '#/components/schemas/Response'
          description: "This response code will be returned in scenarios where the\
            \ inquiry is successful, all possible response codes are explained below\
            \ in the model."
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadFormattedErrorMessage-1'
          description: This response code will be returned in scenarios where the
            inquiry is badly formatted.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RejectedRequestResponse-1'
          description: "This response code will be returned in scenarios where the\
            \ inquiry is rejected, all possible reasons for request rejection are\
            \ explained below in the model."
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorMessage-1'
          description: This response code will be returned in scenarios where the
            inquiry could not be sucessfully processed.
      summary: Acquirer API for PANs Inquiry
      tags:
      - Visa Account Updater Acquirer API
      x-codegen-request-body-name: inquiryRequest
      x-operationVersions:
      - label: v1 - Latest
        operationPointer: '#/paths/~1vau~1acquirer-api~1v1~1inquiry/post'
        default: false
      x-hideTryIt: true
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Visa Account Updater Acquirer API
  /vau/registration-api/v1/callbackconfiguration:
    post:
      description: Push Subscribe Callback API
      operationId: mockAccountUpdateResponse
      parameters:
      - description: VDP MLE keyId used for encryption and decryption
        explode: false
        in: header
        name: keyId
        required: false
        schema:
          type: string
        style: simple
      - description: Correlation ID for tracking requests
        explode: false
        in: header
        name: X-Correlation-ID
        required: false
        schema:
          type: string
        style: simple
      - explode: false
        in: header
        name: Accept
        required: false
        schema:
          enum:
          - application/json
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncryptedPayload'
        required: true
      responses:
        "200":
          content:
            aplications/json:
              schema:
                $ref: '#/components/schemas/OutboundResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundResponse'
          description: This response code will be returned in scenarios where account
            update processing is successful. All possible response codes are explained
            in the model
        "400":
          content:
            aplications/json:
              schema:
                $ref: '#/components/schemas/BadFormattedErrorMessage-2'
            application/json:
              schema:
                $ref: '#/components/schemas/BadFormattedErrorMessage-2'
          description: This response code will be returned in scenarios where the
            request is badly formatted.
        "500":
          content:
            aplications/json:
              schema:
                $ref: '#/components/schemas/GenericErrorMessage'
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorMessage'
          description: This response code will be returned in scenarios where the
            ping test request could not be sucessfully processed.
      summary: Push Subscribe Callback API
      tags:
      - Push Subscribe Callback API
      x-operationVersions:
      - label: v1 - Latest
        operationPointer: '#/paths/~1vau~1registration-api~1v1~1callbackconfiguration/post'
        default: false
      x-hideSample: true
      x-hideTryIt: true
      x-content-type: application/json
      x-accepts:
      - aplications/json
      - application/json
      x-tags:
      - tag: Push Subscribe Callback API
components:
...
  schemas:
...
    GenericErrorMessage:
      description: "[Reject Request]"
      example:
        rejectReason: "Internal Server Error, please try again after sometime"
        rejectReasonCode: E
      properties:
        rejectReason:
          description: |
            This fields contains following reject reasons
            * Internal Server Error, please try again after sometime
          example: "Internal Server Error, please try again after sometime"
          type: string
        rejectReasonCode:
          description: |
            This Fields contains following Reject reason codes:
            * E
          example: E
          type: string
      type: object
    EncryptedPayload:
      example:
        encData: encData
      properties:
        encData:
          type: string
      type: object
...
    stopAdviceRequest:
      description: stopAdviceRequest
      properties:
        stopadvices:
          description: "[List of stop advice]"
          items:
            $ref: '#/components/schemas/stopAdvice'
          type: array
        issuerSegmentId:
          description: "[This field must contain a valid VAU issuer segment id]"
          example: 2
          format: int32
          type: integer
      required:
      - issuerSegmentId
      - stopadvices
      type: object
...
    BadFormattedErrorMessage:
      description: "[Inquiry is badly formatted]"
      example:
        rejectReason: Request is badly formatted and/or the required field(s) is (are)
          missing
        rejectReasonCode: B
      properties:
        rejectReason:
          description: |
            This fields contains following reject reasons
            * Request is badly formatted and/or the required field(s) is (are) missing
          example: Request is badly formatted and/or the required field(s) is (are)
            missing
          type: string
        rejectReasonCode:
          description: |
            This Fields contains following Reject reason codes:
            * B
          example: B
          type: string
      type: object
...
    Request:
      description: "[Acquirer-api request body]"
      example:
        inquiries:
        - cardholderAccountNumber: 400000XX00000000
          expirationDate: "2204"
        - cardholderAccountNumber: 400000XX00000000
          expirationDate: "2204"
        acquirerSegmentId: 2
        merchantId: "000000705853"
        acquirerOrMerchantProprietaryInfo: 180720216995565000113            XXXXAIBR
        subMerchantName: SUB-MERCHANT-NAME
      properties:
        inquiries:
          description: "[List of Inquiries]"
          items:
            $ref: '#/components/schemas/Inquiry'
          type: array
        merchantId:
          description: "[Acquirer assigns the merchant ID. This field must contain\
            \ a valid VAU Merchant ID.]"
          example: "000000705853"
          type: string
        subMerchantName:
          description: |-
            [If inquiry is submitted by a merchant, then this field must be empty.
            If inquiry is submitted by a third-party aggregator on behalf of a merchant, then this field must be populated with the sub-merchant name. Acquirers and processors must pass the value to Visa. No special characters are allowed.]
          example: SUB-MERCHANT-NAME
          type: string
        acquirerSegmentId:
          description: "[This field must contain a valid VAU Acquirer Segment ID.]"
          example: 2
          format: int32
          type: integer
        acquirerOrMerchantProprietaryInfo:
          description: "[If provided, it contains acquirer or merchant proprietary\
            \ information such as a customer ID that is carried in the merchant’s\
            \ back-end system.]"
          example: 180720216995565000113            XXXXAIBR
          type: string
      required:
      - acquirerSegmentId
      - inquiries
      - merchantId
      title: Request
      type: object
...
    RejectedStopAdviceRequestResponse:
      description: "[Reject Request]"
      example:
        issuerSegmentId: 2
        rejectReason: Issuer segment is not enrolled
        rejectReasonCode: A
      properties:
        rejectReason:
          description: |
            [This field contains the description of the reject code:
            * A—Issuer segment is not enrolled
            * T—Max stop advices threshold passed for single request, current threshold is:100
            * V—No data found in stop advice field
          example: Issuer segment is not enrolled
          type: string
        issuerSegmentId:
          description: "[This field must contain a valid VAU Issuer Segment ID.]"
          example: 2
          format: int32
          type: integer
        rejectReasonCode:
          description: |
            [This field contains one of the following codes:
            * A—Issuer segment is not enrolled
            * T—Max stop advices threshold passed for single request, current threshold is:100
            * V—No data found in stop advice field
          example: A
          type: string
      required:
      - issuerSegmentId
      type: object
    StopAdviceResponse:
      description: "[Issuer-stop-advice-api response body]"
      example:
        rejects:
        - cardholderAccountNumber: 400000XX00000000
          cardAcceptorId: "89389898"
          rejectCode: G
          rejectMessage: Account Number contains non-numeric characters or is blank
          startDate: 2000-01-23
          merchantName: TestMerchant
        - cardholderAccountNumber: 400000XX00000000
          cardAcceptorId: "89389898"
          rejectCode: G
          rejectMessage: Account Number contains non-numeric characters or is blank
          startDate: 2000-01-23
          merchantName: TestMerchant
        issuerSegmentId: 2
        responses:
        - cardholderAccountNumber: 400000XX00000000
          cardAcceptorId: "89389898"
          responseMessage: Stop advice applied in both RTVAU and VAU
          startDate: 2000-01-23
          merchantName: TestMerchant
          responseCode: "1"
        - cardholderAccountNumber: 400000XX00000000
          cardAcceptorId: "89389898"
          responseMessage: Stop advice applied in both RTVAU and VAU
          startDate: 2000-01-23
          merchantName: TestMerchant
          responseCode: "1"
      properties:
        rejects:
          description: "[List of rejected stop advice responses]"
          items:
            $ref: '#/components/schemas/RejectedStopAdviceResponse'
          type: array
        responses:
          description: "[List of successful stop advice responses]"
          items:
            $ref: '#/components/schemas/SuccessfulStopAdviceResponse'
          type: array
        issuerSegmentId:
          description: "[This field must contain a valid VAU issuer segment id.]"
          example: 2
          format: int32
          type: integer
      required:
      - issuerSegmentId
      - rejects
      - responses
      type: object
...